backend/pyproject.toml
[project]
name = "elevatria"
version = "1.0.0"
description = "EleVatria - FastAPI + MongoDB Backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # MongoDB
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "beanie>=1.25.0",
    # Pydantic
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Data processing
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.0",
    "pymupdf>=1.24.0",  # Fast PDF parsing (provides fitz)
    "pypdf>=4.0.0",     # PDF text extraction
    # PDF generation & manipulation
    "PyPDF2>=3.0.0",
    "reportlab>=4.0.0",
    # HTTP & streaming
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    # Cache
    "redis>=5.0.0",
    # OpenAI Agents SDK
    "openai>=1.0.0",
    # Environment & auth
    "python-dotenv>=1.0.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    # Utilities
    "loguru>=0.7.0",
    "email-validator>=2.0.0",
    # Fuzzy string matching
    "rapidfuzz>=3.0.0",
    # S3-compatible storage
    "boto3>=1.34.0",
    # Token counting
    "tiktoken>=0.7.0",
    "langchain>=0.2.0, <0.3.0",
    "langchain-community>=0.2.0, <0.3.0",
    "langchain-google-genai>=1.0.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "pypdf>=4.0.0",
    "faster-whisper>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]
# For data migration from PostgreSQL
migration = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
]
# Email providers
email = [
    "resend>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
asyncio_mode = "auto"
